Zdruzil entitetni tip clana v enega
