+++ skavtko/entitete/src/main/java/si/skavtko/entitete/Srecanje.java
@@ -1,8 +1,6 @@
 package si.skavtko.entitete;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
-import java.time.LocalTime;
 import java.util.Set;
 
 import javax.persistence.Basic;
