+++ skavtko/entitete/src/main/java/si/skavtko/entitete/Termin.java
@@ -4,7 +4,6 @@ package si.skavtko.entitete;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.IdClass;
 import javax.persistence.ManyToOne;
 
 public class Termin {
